+++ src/components/common/Auth/AuthSessionManager.tsx
@@ -2,34 +2,14 @@ import { useEffect, useRef } from 'react';
 import type { ReactNode } from 'react';
 import { clearClientSession } from '@apis/auth/sessionActions';
 import { restoreSession } from '@apis/auth/sessionRestore';
-import Error from '@components/common/Error/Error';
 import useAuthStore from '@stores/authStore';
 import useSocketStore from '@stores/socketStore';
-import { HTTP_ERROR_MESSAGE, HTTP_STATUS_CODE } from '@constants/api';
 import { AUTH_RESTORE_DISABLED_KEY } from '@constants/storage';
-import type { AuthStatus, AuthUnavailableReason } from '@type/auth';
+import type { AuthStatus } from '@type/auth';
 
 interface AuthSessionManagerProps {
 	children: ReactNode;
 }
-
-const getAuthErrorMessage = (reason: AuthUnavailableReason | null) => {
-	switch (reason?.type) {
-		case 'network':
-			return HTTP_ERROR_MESSAGE.NETWORK;
-		case 'server':
-			return HTTP_ERROR_MESSAGE[HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR];
-		case 'client':
-		default:
-			return HTTP_ERROR_MESSAGE.DEFAULT;
-	}
-};
-
-const retrySessionRestore = () => {
-	useAuthStore.getState().startBootstrapping();
-
-	return restoreSession();
-};
 
 const useRestoreSessionOnMount = () => {
 	useEffect(() => {
@@ -73,20 +53,10 @@ const useSyncSocketWithAuthStatus = (status: AuthStatus) => {
 
 const AuthSessionManager = ({ children }: AuthSessionManagerProps) => {
 	const status = useAuthStore((state) => state.status);
-	const unavailableReason = useAuthStore((state) => state.unavailableReason);
 
 	useRestoreSessionOnMount();
 	useSyncSessionAcrossTabs();
 	useSyncSocketWithAuthStatus(status);
-
-	if (status === 'unavailable') {
-		return (
-			<Error
-				errorMessage={getAuthErrorMessage(unavailableReason)}
-				resetError={retrySessionRestore}
-			/>
-		);
-	}
 
 	return children;
 };
