+++ src/components/common/Auth/GuestOnlyGuard.tsx
@@ -3,18 +3,18 @@ import UnavailableError from '@components/common/Auth/UnavailableError';
 import useAuthStore from '@stores/authStore';
 import { PATH } from '@constants/path';
 
-const AuthGuard = () => {
+const GuestOnlyGuard = () => {
 	const status = useAuthStore((state) => state.status);
 
 	switch (status) {
 		case 'bootstrapping':
 			return null;
 		case 'guest':
-			return <Navigate to={PATH.SIGNIN} replace />;
+			return <Outlet />;
 		case 'unavailable':
 			return <UnavailableError />;
 		case 'authenticated':
-			return <Outlet />;
+			return <Navigate to={PATH.FEED} replace />;
 		default: {
 			const unhandledStatus: never = status;
 			throw new Error(`처리되지 않은 인증 상태입니다: ${unhandledStatus}`);
@@ -22,4 +22,4 @@ const AuthGuard = () => {
 	}
 };
 
-export default AuthGuard;
+export default GuestOnlyGuard;
